+++ src/models/types.ts
@@ -1,4 +1,0 @@
-export * from './assetModel'
-export * from './portfolio'
-export * from './transactionModel'
-export * from './userModel'